--- src/main/kotlin/com/example/mykku/auth/tool/OauthClient.kt
@@ -1,9 +1,0 @@
-package com.example.mykku.auth.tool
-
-interface OauthClient<T, U> {
-    fun getAuthUrl(): String
-    
-    fun exchangeCodeForToken(code: String): T
-
-    fun getUserInfo(token: String): U
-}
