+++ src/main/kotlin/com/example/mykku/auth/tool/client/GoogleOauthClient.kt
@@ -1,23 +1,23 @@
-package com.example.mykku.auth.tool
+package com.example.mykku.auth.tool.client
 
-import com.example.mykku.auth.dto.KakaoUserInfo
+import com.example.mykku.auth.dto.GoogleUserInfo
 import com.example.mykku.auth.exception.AuthException
 import org.springframework.http.HttpHeaders
 import org.springframework.stereotype.Component
 import org.springframework.web.client.RestClient
 
 @Component
-class KakaoOauthClient(
+class GoogleOauthClient(
     private val restClient: RestClient
 ) : AbstractOauthClient() {
 
-    fun verifyAndGetUserInfo(accessToken: String): KakaoUserInfo {
-        return executeOauthRequest("KAKAO") {
+    fun verifyAndGetUserInfo(accessToken: String): GoogleUserInfo {
+        return executeOauthRequest("GOOGLE") {
             restClient.get()
-                .uri("https://kapi.kakao.com/v2/user/me")
+                .uri("https://www.googleapis.com/oauth2/v2/userinfo")
                 .header(HttpHeaders.AUTHORIZATION, "Bearer $accessToken")
                 .retrieve()
-                .body(KakaoUserInfo::class.java)
+                .body(GoogleUserInfo::class.java)
                 ?: throw AuthException.oauthUserInfoFailed()
         }
     }
