+++ src/main/kotlin/com/example/mykku/auth/tool/client/KakaoOauthClient.kt
@@ -1,23 +1,23 @@
-package com.example.mykku.auth.tool
+package com.example.mykku.auth.tool.client
 
-import com.example.mykku.auth.dto.GoogleUserInfo
+import com.example.mykku.auth.dto.KakaoUserInfo
 import com.example.mykku.auth.exception.AuthException
 import org.springframework.http.HttpHeaders
 import org.springframework.stereotype.Component
 import org.springframework.web.client.RestClient
 
 @Component
-class GoogleOauthClient(
+class KakaoOauthClient(
     private val restClient: RestClient
 ) : AbstractOauthClient() {
 
-    fun verifyAndGetUserInfo(accessToken: String): GoogleUserInfo {
-        return executeOauthRequest("GOOGLE") {
+    fun verifyAndGetUserInfo(accessToken: String): KakaoUserInfo {
+        return executeOauthRequest("KAKAO") {
             restClient.get()
-                .uri("https://www.googleapis.com/oauth2/v2/userinfo")
+                .uri("https://kapi.kakao.com/v2/user/me")
                 .header(HttpHeaders.AUTHORIZATION, "Bearer $accessToken")
                 .retrieve()
-                .body(GoogleUserInfo::class.java)
+                .body(KakaoUserInfo::class.java)
                 ?: throw AuthException.oauthUserInfoFailed()
         }
     }
