+++ src/main/kotlin/com/example/mykku/auth/tool/strategy/AppleLoginStrategy.kt
@@ -1,6 +1,11 @@
-package com.example.mykku.auth.tool
+package com.example.mykku.auth.tool.strategy
 
 import com.example.mykku.auth.dto.LoginResponse
+import com.example.mykku.auth.tool.JwtTokenProvider
+import com.example.mykku.auth.tool.MemberOrchestrator
+import com.example.mykku.auth.tool.OAuthLoginStrategy
+import com.example.mykku.auth.tool.OAuthMemberExtractor
+import com.example.mykku.auth.tool.client.AppleOauthClient
 import com.example.mykku.member.domain.SocialProvider
 import org.springframework.stereotype.Component
