+++ src/main/kotlin/com/example/mykku/auth/tool/strategy/GoogleLoginStrategy.kt
@@ -1,24 +1,29 @@
-package com.example.mykku.auth.tool
+package com.example.mykku.auth.tool.strategy
 
 import com.example.mykku.auth.dto.LoginResponse
+import com.example.mykku.auth.tool.JwtTokenProvider
+import com.example.mykku.auth.tool.MemberOrchestrator
+import com.example.mykku.auth.tool.OAuthLoginStrategy
+import com.example.mykku.auth.tool.OAuthMemberExtractor
+import com.example.mykku.auth.tool.client.GoogleOauthClient
 import com.example.mykku.member.domain.SocialProvider
 import org.springframework.stereotype.Component
 
 @Component
-class KakaoLoginStrategy(
-    private val kakaoOauthClient: KakaoOauthClient,
+class GoogleLoginStrategy(
+    private val googleOauthClient: GoogleOauthClient,
     private val oauthMemberExtractor: OAuthMemberExtractor,
     private val memberOrchestrator: MemberOrchestrator,
     private val jwtTokenProvider: JwtTokenProvider
 ) : OAuthLoginStrategy {
 
     override fun supports(provider: SocialProvider): Boolean {
-        return provider == SocialProvider.KAKAO
+        return provider == SocialProvider.GOOGLE
     }
 
     override fun login(token: String): LoginResponse {
-        val userInfo = kakaoOauthClient.verifyAndGetUserInfo(token)
-        val memberInfo = oauthMemberExtractor.extractFromKakao(userInfo)
+        val userInfo = googleOauthClient.verifyAndGetUserInfo(token)
+        val memberInfo = oauthMemberExtractor.extractFromGoogle(userInfo)
         val (member, isExistingUser) = memberOrchestrator.findOrCreate(memberInfo)
         return jwtTokenProvider.createLoginResponse(member, memberInfo.email, isExistingUser)
     }
