+++ src/test/kotlin/com/example/mykku/auth/tool/client/NaverOauthClientTest.kt
@@ -1,10 +1,10 @@
-package com.example.mykku.auth.tool
+package com.example.mykku.auth.tool.client
 
 import com.example.mykku.BaseToolTest
 import com.example.mykku.auth.dto.NaverUserInfo
 import com.example.mykku.auth.dto.NaverUserResponse
-import com.example.mykku.auth.exception.AuthException
 import com.example.mykku.auth.exception.AuthErrorCode
+import com.example.mykku.auth.exception.AuthException
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
 import org.mockito.InjectMocks
